Add extra_options to search
My use case is just to add defType=edismax which means I get better
results.  The way to use it is to do:

    myquery = myquery.add_extra(defType='edismax')

And you could also use it to add other dismax parameters (eg qf, mm ...)

There are probably better ways to do this, and it is certainly unsafe
with zero error checking, but it allows me to do what I need.